split SBuiltinTest
